requirements.txt
Django==2.1
django-geoposition==0.3.0
pkg-resources==0.0.0
pytz==2018.5
